Move test config to its own module, instead of constants.py
