+++ tests/aws/services/cloudformation/resources/test_cdk.py
@@ -6,8 +6,8 @@ import requests
 from localstack_snapshot.snapshots.transformer import SortingTransformer
 
 from localstack import config
-from localstack.constants import TEST_AWS_ACCESS_KEY_ID
 from localstack.testing.pytest import markers
+from localstack.testing.test_config import TEST_AWS_ACCESS_KEY_ID
 from localstack.utils.aws.request_context import mock_aws_request_headers
 from localstack.utils.files import load_file
 from localstack.utils.strings import short_uid
