+++ tests/aws/services/cloudwatch/test_cloudwatch.py
@@ -13,11 +13,11 @@ import pytest
 import requests
 
 from localstack import config
-from localstack.constants import TEST_AWS_ACCESS_KEY_ID
 from localstack.services.cloudwatch.provider import PATH_GET_RAW_METRICS
 from localstack.testing.aws.util import is_aws_cloud
 from localstack.testing.pytest import markers
 from localstack.testing.snapshots.transformer_utility import TransformerUtility
+from localstack.testing.test_config import TEST_AWS_ACCESS_KEY_ID
 from localstack.utils.aws import arns
 from localstack.utils.aws.request_context import mock_aws_request_headers
 from localstack.utils.common import retry, short_uid, to_str
