+++ tests/aws/services/logs/test_logs.py
@@ -8,8 +8,9 @@ from localstack_snapshot.pytest.snapshot import is_aws
 from localstack_snapshot.snapshots.transformer import KeyValueBasedTransformer
 
 from localstack.aws.api.lambda_ import Runtime
-from localstack.constants import APPLICATION_AMZ_JSON_1_1, TEST_AWS_REGION_NAME
+from localstack.constants import APPLICATION_AMZ_JSON_1_1
 from localstack.testing.pytest import markers
+from localstack.testing.test_config import TEST_AWS_REGION_NAME
 from localstack.utils import testutil
 from localstack.utils.aws import arns
 from localstack.utils.common import now_utc, poll_condition, retry, short_uid
