+++ tests/aws/services/stepfunctions/v2/test_stepfunctions_v2.py
@@ -4,13 +4,13 @@ import os
 
 import pytest
 
-from localstack.constants import (
-    SECONDARY_TEST_AWS_ACCESS_KEY_ID,
-    SECONDARY_TEST_AWS_SECRET_ACCESS_KEY,
-)
 from localstack.services.events.v1.provider import TEST_EVENTS_CACHE
 from localstack.testing.aws.util import is_aws_cloud
 from localstack.testing.pytest import markers
+from localstack.testing.test_config import (
+    SECONDARY_TEST_AWS_ACCESS_KEY_ID,
+    SECONDARY_TEST_AWS_SECRET_ACCESS_KEY,
+)
 from localstack.utils import testutil
 from localstack.utils.aws import arns
 from localstack.utils.files import load_file
