+++ tests/unit/services/opensearch/test_cluster_manager.py
@@ -2,8 +2,8 @@ import pytest
 
 from localstack import config
 from localstack.aws.api.opensearch import EngineType
-from localstack.constants import TEST_AWS_ACCOUNT_ID
 from localstack.services.opensearch.cluster_manager import DomainKey, build_cluster_endpoint
+from localstack.testing.test_config import TEST_AWS_ACCOUNT_ID
 
 
 class TestBuildClusterEndpoint:
